move all types back to types.ts file
