--- src/types/index.ts
@@ -1,6 +1,0 @@
-export * from './base';
-export * from './event';
-export * from './filter';
-export * from './option';
-export * from './response';
-export * from './util';
